airflow_docker/airflow_docker/requirements_airflow.txt
snowflake-connector-python
pymongo
pandas